maintenance/improvement * fix userid casting to int   never use bindParam with an integer variable without PDO::PARAM_INT or the variable will be casted to string,   better use bindValue to avoid this confusion * remove bindValue()/bindParam() default parameter PDO::PARAM_STR where possible * less casting * move getNextCustomId to AbstractConcreteEntity * add some missing use statements

+++ src/commands/ExportResources.php
@@ -48,7 +48,7 @@ class ExportResources extends Command
     {
         $categoryId = (int) $input->getArgument('category_id');
         $userid = (int) $input->getArgument('userid');
-        $teamid = (int) (new UsersHelper($userid))->getTeamsFromUserid()[0]['id'];
+        $teamid = (new UsersHelper($userid))->getTeamsFromUserid()[0]['id'];
         $absolutePath = $this->Fs->getPath(sprintf(
             'export-%s-category_id-%d.eln',
             date('Y-m-d_H-i-s'),
